chore(deps): update angular and related to 1.3.x
Angular 1.3.x has many improvements. The driving factor in upgrading
is the potential to improve rendering speed of repeaters.

rc.0 introduced some changes:

* baseHref is required but not available in mocks. This is an angular
bug. Work around it by decorating the $browser service to supply it.
* It seems to require html for directive templates even if they aren't
needed by the tests. Respond with actual directive template html when
it is requested to keep angular from throwing an error.