java: convenience .sh script for full gradle run
cleanstart.sh downloads the current seed data and does a full, clean
setup of the Samplestack Java tier and finishes by launching bootrun.

It is not expected to work on Windows. In the future, a cross-platform
solution may be developed in gradle.

For now, it represents the expected path for successful launch of the
Java stack after retrieving the code from git (minus all browser-related
commands).

It should work for a clean install and for a reinstall.