e2e: work around cucumber next callback -> notifyOk
Cucumber's next function isn't promisified and thus considers a
truthy parameter to be an indicator of failure.

This is not an issue when we are dealing with chai-as-promised
expectations because they have .and.notify support.

Create a sugar function to be used when we don't have an expectation so
that any value which happens to be resolved in by a steps function
does not trick Cucumber into thinking something went wrong.