browser: date picker validation and interactions for manual entry
Adds on blur and on enter key as trigger to apply picker date value
while manually entering a date.

Add validation where if the currently focused date input is set to an
invalid range, the other input is set to match its date, setting
the range to a span of one month.

Closes #412 and Closes #414