browser: CORS/CSRF
- make middle-tier REST server address configurable and configure it
to use the middle-tier on port 8090 directly, skipping the proxy, thus
using CORS. use port 8090/CORS for REST services
- enable CSRF mode in the browser
- only seek a token on GET /v1/session
- when middle-tier returns initial /v1/session there is garbage in the
response. DO throw it away (key on whether username is present).
DO NOT throw away responses where there is a username -- these are
legitimate restorations/refetches of session.
- Only get a CSRF token when needed (POST /v1/session).  Restore from
cookie if available when session restoring
- wipe out *all* cookies on logout. Don't just remove sessionId, remove
CSRF, too, so that we'll get a new token if/when we try to login again.
- patch unit tests for new CSRF algorithms
- no matter what goes wrong during logout, delete cookies
- Don't use a potentially old CSRF token from the cookie in cases
where the user is actually trying to POST their credentials.
-  avoid $cookieStore trying to parse tokens as JSON, always use
$cookies
- httpInterceptor: clear `outstanding` variable when csrf token request
is done
- delete all cookie values if they are inconsistent with token
presence/absence
- only worry about REST calls, not static
- set CSRF token whenever it's in a response. Ensures that the latest
token known to the server is the one we're using.
- only write to session object if it's a login-type response
- If a 401/403 happens, trigger a restoreSession.
This isn't optimal, but give the lack of detail in responses from the
server for credentials/csrf conditions, for now should allow the
browser to "clean itself up" to the point where it at least demonstrates
to the user that they aren't being treated as logged in as well as
restore the proper cookie/headers situation.

Closes #40, #41