browser: sort-of enable CSRF
This commit sort-of enables CSRF, but it doesn't look like the middle
tier is really protecting itself.

Additionally, it seems that the middle tier cannot run both CORS and
CSRF simultaneously because the CORS protection disallows the
X-CSRF-TOKEN header.

TODO for the browser is to put the token in its session cookie, since
refreshes would otherwise break it (if the CSRF protection was actually
working).