browser: datepicker clear btw in pop-up fix
Fixes #469 issue with clean button in datepicker dialogue no updating the search criteria, but just clearing the field.

Attempted to add a e2e test, but was unable to get sendKeys() to work, for whatever reason.  So I'll need to revisit this test stub when there's more time.  For now it is marked as @broken.