enable java middle-tier with "live browser" and e2e
Re-enable use of java middle-tier with the "live" version of the
browser on port 3000, rather than only supporting static build on port
8090.

Allows for use of Java middle tier while also developing or testing
the source code of the browser. This works for e2e as well as run/watch
modes.

To use: supply the `--middle-tier=java` or `middle-tier=external` flag.
(The `-middle-tier=java` flag is different from `external` only under
`e2e`, where if --middle-tier=java` the automation will actually drive
the middle-tier to be launched, whereas with `external`, it is assumed
that the developer has launched the java middle-tier in a separate
terminal.

Closes #603